seperated components into diff files
