+++ public/locale/en.json
@@ -341,7 +341,6 @@
   "abandoned": "Abandoned",
   "abnormal": "Abnormal",
   "abnormal_coded_value_set": "Valueset - Abnormal",
-  "abnormal_only": "Abnormal only",
   "aborted": "Aborted",
   "absolute_date": "Absolute Date",
   "accept": "Accept",
@@ -3674,7 +3673,6 @@
   "next_week": "Next Week",
   "next_week_short": "Next wk",
   "no": "No",
-  "no_abnormal_readings": "No abnormal readings",
   "no_account_found": "No account found",
   "no_accounts_found": "No accounts found",
   "no_active_account_found": "No active account found",
@@ -3909,6 +3907,7 @@
   "no_permission_to_view_clinical_data_description": "Contact your administrator to grant you access.",
   "no_permission_to_view_page": "You do not have permissions to view this page",
   "no_plots_configured": "No plots configured",
+  "no_plottable_values": "No plottable values",
   "no_policy_added": "No Insurance Policy Added",
   "no_policy_found": "No Insurance Policy Found for this Patient",
   "no_practitioners_found": "No practitioners found",
@@ -4920,7 +4919,6 @@
   "recordings_count": "{{count}} recordings",
   "recurrence": "Recurrence",
   "redirected_to_create_consultation": "Note: You will be redirected to create consultation form. Please complete the form to finish the transfer process",
-  "ref": "Ref",
   "reference": "Reference",
   "reference_details": "Reference Details",
   "reference_no": "Reference No",
@@ -5326,8 +5324,6 @@
   "scheduling": "Scheduling",
   "scribe__reviewing_field": "Reviewing field {{currentField}} / {{totalFields}}",
   "scribe_error": "Could not autofill fields",
-  "scroll_left": "Scroll left",
-  "scroll_right": "Scroll right",
   "scroll_to_question": "Scroll to question",
   "scroll_to_top": "Scroll to top",
   "search": "Search",
@@ -5871,7 +5867,6 @@
   "show_status": "Show Status",
   "show_unread_notifications": "Show Unread",
   "showing": "Showing",
-  "showing_abnormal_readings": "Showing abnormal readings",
   "showing_all_appointments": "Showing all appointments",
   "showing_all_medications": "Showing all medications",
   "showing_all_nutritional_products": "Showing all nutritional products",
